+++ contracts/staking/src/contract.rs
@@ -458,13 +458,12 @@ mod tests {
 
     fn sample_delegation<U: Into<HumanAddr>>(addr: U, amount: Coin) -> FullDelegation {
         let can_redelegate = amount.clone();
-        let accumulated_rewards = coin(0, &amount.denom);
         FullDelegation {
             validator: addr.into(),
             delegator: HumanAddr::from(MOCK_CONTRACT_ADDR),
             amount,
             can_redelegate,
-            accumulated_rewards,
+            accumulated_rewards: Vec::new(),
         }
     }